+++ darr/tests/test_array.py
@@ -5,7 +5,7 @@ import shutil
 
 import numpy as np
 
-from darr.array import asarray, create_array, create_basedatadir, Array, \
+from darr.array import asarray, create_array, create_datadir, Array, \
     numtypesdescr, truncate_array, delete_array, AppendDataError, \
     numtypedescriptiontxt
 from .utils import tempdir, tempdirfile
@@ -295,6 +295,12 @@ class TestArray(DarrTestCase):
         self.assertArrayIdentical(self.tempar[2:4],
                                   np.array([1, 1], dtype=self.tempar.dtype))
 
+    def test_datadirexistence(self):
+        with tempdirfile() as filename:
+            dar = create_array(path=filename, shape=(2,), fill=0,
+                               dtype='int64', overwrite=True)
+            self.assertEqual(filename, dar.datadir.path)
+
     def test_str(self):
         with tempdirfile() as filename:
             dar = create_array(path=filename, shape=(2,), fill=0,
@@ -355,7 +361,7 @@ class TestReadArrayDescr(DarrTestCase):
             arrayinfo = dar._arrayinfo.copy()
             vs = f"1{arrayinfo['darrversion']}"
             arrayinfo['darrversion'] = vs
-            dar._write_jsondict(dar._arraydescrfilename, arrayinfo,
+            dar._datadir._write_jsondict(dar._arraydescrfilename, arrayinfo,
                                 overwrite=True)
             self.assertWarns(UserWarning, Array, dar.path)
 
@@ -365,7 +371,7 @@ class TestReadArrayDescr(DarrTestCase):
                                dtype='int64', overwrite=True)
             arrayinfo = dar._arrayinfo.copy()
             arrayinfo['shape'] = ['a', 3]
-            dar._write_jsondict(dar._arraydescrfilename, arrayinfo,
+            dar._datadir._write_jsondict(dar._arraydescrfilename, arrayinfo,
                                 overwrite=True)
             self.assertRaises(TypeError, Array, dar.path)
 
@@ -375,11 +381,11 @@ class TestReadArrayDescr(DarrTestCase):
                                dtype='int64', overwrite=True)
             arrayinfo = dar._arrayinfo.copy()
             arrayinfo['arrayorder'] = 'D'
-            dar._write_jsondict(dar._arraydescrfilename, arrayinfo,
+            dar._datadir._write_jsondict(dar._arraydescrfilename, arrayinfo,
                                 overwrite=True)
             self.assertRaises(ValueError, Array, dar.path)
             arrayinfo['arrayorder'] = '[D]'
-            dar._write_jsondict(dar._arraydescrfilename, arrayinfo,
+            dar._datadir._write_jsondict(dar._arraydescrfilename, arrayinfo,
                                 overwrite=True)
             self.assertRaises(Exception, Array, dar.path)
 
@@ -387,7 +393,7 @@ class TestReadArrayDescr(DarrTestCase):
         with tempdirfile() as filename:
             dar = create_array(path=filename, shape=(2,4), fill=0,
                                dtype='int64', overwrite=True)
-            dar._update_jsondict(dar._arraydescrpath.absolute(),
+            dar._datadir._update_jsondict(dar._arraydescrpath.absolute(),
                                  {'arrayorder': 'F'})
             dar = Array(filename)
             self.assertIn("Column-major", numtypedescriptiontxt(dar))
@@ -396,7 +402,7 @@ class TestReadArrayDescr(DarrTestCase):
         with tempdirfile() as filename1, tempdirfile() as filename2:
             dar = create_array(path=filename1, shape=(2, 4), fill=0,
                                dtype='int64', overwrite=True)
-            dar._update_jsondict(dar._arraydescrpath.absolute(),
+            dar._datadir._update_jsondict(dar._arraydescrpath.absolute(),
                                  {'arrayorder': 'F'})
             dar = Array(filename1)
             self.assertWarns(UserWarning, asarray, path=filename2, array=dar,
@@ -449,7 +455,7 @@ class TestConsistency(DarrTestCase):
                                dtype='int64', overwrite=True)
             arrayinfo = dar._arrayinfo.copy()
             arrayinfo['numtype'] = 'int32'
-            dar._write_jsondict(dar._arraydescrfilename, arrayinfo,
+            dar._datadir._write_jsondict(dar._arraydescrfilename, arrayinfo,
                                 overwrite=True)
             self.assertRaises(ValueError, dar._check_arrayinfoconsistency)
             self.assertRaises(ValueError, Array, dar.path)
@@ -648,7 +654,7 @@ class TestOpenFile(DarrTestCase):
         with tempdir() as dirname:
             dar = create_array(path=dirname, shape=(0, 2), dtype='int64',
                                overwrite=True, accessmode='r+')
-            with dar.open_file('notes.txt', 'a') as f:
+            with dar._datadir.open_file('notes.txt', 'a') as f:
                 f.write('test\n')
             path = dar.path / 'notes.txt'
             self.assertEqual(path.exists(), True)
@@ -657,9 +663,9 @@ class TestOpenFile(DarrTestCase):
         with tempdir() as dirname:
             dar = create_array(path=dirname, shape=(0, 2), dtype='int64',
                                overwrite=True, accessmode='r+')
-            for fn in dar._filenames:
+            for fn in dar._protectedfiles:
                 with self.assertRaises(OSError):
-                    with dar.open_file(fn, 'a') as f:
+                    with dar._datadir.open_file(fn, 'a') as f:
                         f.write('test\n')
 
 
@@ -688,7 +694,7 @@ class TruncateData(DarrTestCase):
 
     def test_donottruncatenondarrdir(self):
         with tempdirfile() as filename:
-            bd = create_basedatadir(filename)
+            bd = create_datadir(filename)
             bd._write_jsondict('test.json', {'a': 1})
             self.assertRaises(TypeError, truncate_array, filename, 3)
 
@@ -732,14 +738,14 @@ class DeleteArray(DarrTestCase):
     def test_donotdeletenondarrfile(self):
         with tempdirfile() as filename:
             dar = create_array(path=filename, shape=(0, 2), dtype='int64')
-            dar._write_jsondict('test.json', {'a': 1})
+            dar._datadir._write_jsondict('test.json', {'a': 1})
             testpath = dar._path.joinpath('test.json')
             self.assertRaises(OSError, delete_array, dar)
             self.assertEqual(testpath.exists(), True)
 
     def test_donotdeletenondarrdir(self):
         with tempdirfile() as filename:
-            bd = create_basedatadir(filename, overwrite=True)
+            bd = create_datadir(filename, overwrite=True)
             self.assertRaises(TypeError, delete_array, filename)
             bd._write_jsondict('test.json', {'a': 1})
             self.assertRaises(TypeError, delete_array, filename)